Fixed Scrollbars not calculating their slider size properly based on the adjustment values, fixed ScrolledWindow sometimes displaying 2 scrollbars instead of only 1 when both are set to automatic and the child viewport is expanded, fixed OpenGL example not spinning as fast as intended because the clock was restarted twice per frame, fixed a few undefined behavior errors as reported by clang sanitizers, added Wsign-conversion to prevent sign conversion errors, changed to include libELL headers as system headers so that pragmas are no longer required to silence warnings.

+++ examples/Canvas.cpp
@@ -125,7 +125,7 @@ int main() {
 		glMatrixMode( GL_MODELVIEW );
 		glPopMatrix();
 
-		glViewport( 0, 0, app_window.getSize().x, app_window.getSize().y );
+		glViewport( 0, 0, static_cast<int>( app_window.getSize().x ), static_cast<int>( app_window.getSize().y ) );
 
 		glEnable( GL_TEXTURE_2D );
 		glDisable( GL_DEPTH_TEST );
